\name{a}
\docType{methods}
\alias{a}
\alias{get.a}
\alias{get.a,crp.CSFP-method}
\alias{a-methods}
\alias{a,crp.CSFP-method}
\title{Get the parameter \code{a} of the model.}
\description{
The method returns the value of \code{a} (used for calculation of loss distribution) as numeric.
}
\keyword{methods}
