\name{alpha.crp}
\docType{methods}
\alias{alpha.crp}
\alias{get.ALPHA.crp}
\alias{get.ALPHA.crp,crp.CSFP-method}
\alias{alpha.crp-methods}
\alias{alpha.crp,crp.CSFP-method}
\title{Get the maximum cdf levels for VaR of the model.}
\description{
The method returns the value of \code{alpha.crp} as numeric. These are different from the desired CDF levels \code{alpha} because the CDF is not continous.
}
\keyword{methods}
