\name{alpha.max}
\docType{methods}
\alias{alpha.max}
\alias{get.ALPHA.MAX}
\alias{get.ALPHA.MAX,crp.CSFP-method}
\alias{alpha.max-methods}
\alias{alpha.max,crp.CSFP-method}
\title{Get the maximum cdf level for loss distribution}
\description{
 The method returns the value of \code{alpha.max} as numeric.
}
\keyword{methods}
