\name{B}
\docType{methods}
\alias{B}
\alias{get.B}
\alias{get.B,crp.CSFP-method}
\alias{B-methods}
\alias{B,crp.CSFP-method}
\title{Get the parameter \code{B} of the model.}
\description{
The method returns the value of \code{B} (used for calculation of loss distribution) as matrix.
}
\keyword{methods}
