\name{calc.rc}
\docType{methods}
\alias{calc.rc}
\alias{get.CALC.RISK.CONT}
\alias{get.CALC.RISK.CONT,crp.CSFP-method}
\alias{calc.rc-methods}
\alias{calc.rc,crp.CSFP-method}
\title{Set the state of \code{calc.rc}}
\description{
The method changes the models value of \code{calc.rc}.}
\keyword{methods}