\name{calc.rc<-}
\docType{methods}
\alias{calc.rc<-}
\alias{set.CALC.RISK.CONT<-}
\alias{set.CALC.RISK.CONT<-,crp.CSFP,logical-method}
\alias{calc.rc<--methods}
\alias{calc.rc<-,crp.CSFP,logical-method}
\title{Get the value of the \code{calc.rc}}
\description{
The method returns the value of \code{calc.rc} as logical.
}
\keyword{methods}
