\name{changes.calc.portfolio.statistics}
\docType{methods}
\alias{changes.calc.portfolio.statistics}
\alias{get.changes.crp.calc.portfolio.statistics}
\alias{get.changes.crp.calc.portfolio.statistics,crp.CSFP-method}
\alias{changes.calc.portfolio.statistics-methods}
\alias{changes.calc.portfolio.statistics,crp.CSFP-method}
\title{Get the state of \code{changes.calc.portfolio.statistics}.}
\description{
The method returns the value of \code{changes.calc.portfolio.statistics}.}
\keyword{methods}
