\name{changes.export}
\docType{methods}
\alias{changes.export}
\alias{get.changes.crp.export}
\alias{get.changes.crp.export,crp.CSFP-method}
\alias{changes.export-methods}
\alias{changes.export,crp.CSFP-method}
\title{Get the state of \code{changes.export}}
\description{
The method returns the value of \code{changes.export}.}
\keyword{methods}