\name{changes.loss}
\docType{methods}
\alias{changes.loss}
\alias{get.changes.crp.CSFP.loss}
\alias{get.changes.crp.CSFP.loss,crp.CSFP-method}
\alias{changes.loss-methods}
\alias{changes.loss,crp.CSFP-method}
\title{Get the state of \code{changes.loss}}
\description{
The method returns the value of \code{changes.loss}.}
\keyword{methods}
