\name{changes.measure}
\docType{methods}
\alias{changes.measure}
\alias{get.changes.crp.measure}
\alias{get.changes.crp.measure,crp.CSFP-method}
\alias{changes.measure-methods}
\alias{changes.measure,crp.CSFP-method}
\title{Get the state of \code{changes.measure}}
\description{
The method returns the value of \code{changes.measure}.}
\keyword{methods}

