\name{changes.plausi}
\docType{methods}
\alias{changes.plausi}
\alias{get.changes.crp.plausi}
\alias{get.changes.crp.plausi,crp.CSFP-method}
\alias{changes.plausi-methods}
\alias{changes.plausi,crp.CSFP-method}
\title{Get the state of \code{changes.plausi}}
\description{
The method returns the value of \code{changes.plausi}.}
\keyword{methods}
