\name{changes.plot}
\docType{methods}
\alias{changes.plot}
\alias{get.changes.crp.plot}
\alias{get.changes.crp.plot,crp.CSFP-method}
\alias{changes.plot-methods}
\alias{changes.plot,crp.CSFP-method}
\title{Get the state of \code{changes.plot}}
\description{
The method returns the value of \code{changes.plot}.}
\keyword{methods}
