\name{changes.rc.sd}
\docType{methods}
\alias{changes.rc.sd}
\alias{get.changes.crp.CSFP.rc.sd}
\alias{get.changes.crp.CSFP.rc.sd,crp.CSFP-method}
\alias{changes.rc.sd-methods}
\alias{changes.rc.sd,crp.CSFP-method}
\title{Get the state of \code{changes.rc.sd}}
\description{
The method returns the value of \code{changes.rc.sd}.}
\keyword{methods}
