\name{changes.rc.vares}
\docType{methods}
\alias{changes.rc.vares}
\alias{get.changes.crp.CSFP.rc.vares}
\alias{get.changes.crp.CSFP.rc.vares,crp.CSFP-method}
\alias{changes.rc.vares-methods}
\alias{changes.rc.vares,crp.CSFP-method}
\title{Get the state of \code{changes.rc.vares}}
\description{
The method returns the value of \code{changes.rc.vares}.}
\keyword{methods}
