\name{changes.read}
\docType{methods}
\alias{changes.read}
\alias{get.changes.crp.read}
\alias{get.changes.crp.read,crp.CSFP-method}
\alias{changes.read-methods}
\alias{changes.read,crp.CSFP-method}
\title{Get the state of \code{changes.read}}
\description{
The method returns the value of \code{changes.read}.}
\keyword{methods}