\name{CP.NR}
\docType{methods}
\alias{CP.NR}
\alias{get.CP.NR}
\alias{get.CP.NR,crp.CSFP-method}
\alias{CP.NR-methods}
\alias{CP.NR,crp.CSFP-method}
\title{Get the ID numbers of the counterparties of the model}
\description{
The method returns the value of \code{CP.NR} as numeric.}
\keyword{methods}