\name{CP.rating}
\docType{methods}
\alias{CP.rating}
\alias{get.CP.RATING}
\alias{get.CP.RATING,crp.CSFP-method}
\alias{CP.rating-methods}
\alias{CP.rating,crp.CSFP-method}
\title{Get counterparties ratings}
\description{
The method returns the value of \code{CP.rating} as numeric.}
\keyword{methods}
