\name{crp.CSFP.rc.sd}
\docType{methods}
\alias{crp.CSFP.rc.sd}
\alias{crp.CSFP.rc.sd,crp.CSFP-method}
\title{Calculating risk contributions to standard deviation}
\description{
This method calculates the contributions on counterparty level to the portfolio standard deviation. No data out of \code{crp.loss.CSFP} are required.
}
\seealso{
\code{\link{crp.CSFP.rc.vares}},
\code{\link{crp.export}},
\code{\link{crp.CSFP-class}},
\code{\link{crp.CSFP}},
}
\keyword{methods}