\name{crp.export}
\docType{methods}
\alias{crp.export}
\alias{crp.export,crp.CSFP-method}
\title{Export risk contributions}
\description{
This method exports the risk contributions, calculated in \code{crp.CSFP.rc.vares} and \code{crp.CSFP.rc.sd} to a file named "RC.csv" in the \code{PATH.OUT/NAME} directory. In \code{save.memory} mode, the content of the risk contribution attributes will be deleted afterwards.
}
\seealso{
\code{\link{crp.CSFP.rc.vares}},
\code{\link{crp.CSFP.rc.sd}},
\code{\link{crp.CSFP}},
\code{\link{crp.CSFP-class}},
}
\keyword{methods}
