\name{crp.plot}
\docType{methods}
\alias{crp.plot}
\alias{crp.plot,crp.CSFP-method}
\title{Plotting the PDF}
\description{
This method creates a plot in a new window of the loss distribution together with lines indicating the \code{EL.crp}, \code{VaR} and \code{ES}. The X-/Y- ranges are taken from \code{PLOT.RANGE.X /-Y}. A picture of type "jpg" is saved at the output directory.
}
\seealso{
\code{\link{crp.CSFP}},
\code{\link{crp.CSFP-class}},
}
\keyword{methods}