\name{crp.round}
\alias{crp.round}
\title{Rounding numerical values}
\description{
The function rounds numerical values to the next integer. If the first digit after the decimal point is equal to 5 the function rounds up.}
\usage{
crp.round(a)
}
\arguments{
  \item{a}{
A numerical value which should be rounded.}
}
\value{
numeric
}
\author{
Dr. Matthias Fischer}
