\name{crp.summary}
\docType{methods}
\alias{crp.summary}
\alias{crp.summary,crp.CSFP-method}

\title{Summarize portfolio key numbers}
\description{
This method summarizes the input parameter, portfolio statistics from \code{crp.calc.portfolio.statistics} and the results of \code{crp.measure} in a list.
}
\seealso{
\code{\link{crp.write.summary}},
\code{\link{crp.CSFP}},
\code{\link{crp.CSFP-class}},
}
\keyword{methods}
