\name{EL.crp}
\docType{methods}
\alias{EL.crp}
\alias{get.EL.crp}
\alias{get.EL.crp,crp.CSFP-method}
\alias{EL.crp-methods}
\alias{EL.crp,crp.CSFP-method}
\title{Get the expected loss of the model after discretization.}
\description{
The method returns the value of \code{EL.crp} as numeric.}
\keyword{methods}
