\name{ES.cont}
\docType{methods}
\alias{ES.cont}
\alias{get.ES.CONT}
\alias{get.ES.CONT,crp.CSFP-method}
\alias{ES.cont-methods}
\alias{ES.cont,crp.CSFP-method}
\title{Get the expected shortfall contributions}
\description{
The method returns the value of \code{ES.cont} as numeric.}
\keyword{methods}
