\name{ES.tau.cont}
\docType{methods}
\alias{ES.tau.cont}
\alias{get.ES.TAU.CONT}
\alias{get.ES.TAU.CONT,crp.CSFP-method}
\alias{ES.tau.cont-methods}
\alias{ES.tau.cont,crp.CSFP-method}
\title{Get the corresponding tau for expected shortfall contributions}
\description{
The method returns the value of \code{ES.tau.cont} as numeric.}
\keyword{methods}
