\name{file.format}
\docType{methods}
\alias{file.format}
\alias{get.file.format}
\alias{get.file.format,crp.CSFP-method}
\alias{file.format-methods}
\alias{file.format,crp.CSFP-method}
\title{Get the file format of the model}
\description{
The method returns the value of \code{file.format} as character.}
\keyword{methods}