\name{file.format<-}
\docType{methods}
\alias{file.format<-}
\alias{set.file.format<-}
\alias{set.file.format<-,crp.CSFP,character-method}
\alias{file.format<--methods}
\alias{file.format<-,crp.CSFP,character-method}
\title{Set the file format}
\description{
The method changes the models value of \code{file.format}}
\keyword{methods}
