\name{fo}
\alias{fo}
\title{
Function to convert numerical output.}
\description{
The Function writes numerical values as a multiple of the next power of 1000 with respect to three digits in order to make large values more readable. The trailed abbreviation corresponds to 1e3 (Thd), 1e6 (Mio), 1e9 (Bil), 1e12 (Tril).}
\usage{
fo(x)
}
\arguments{
  \item{x}{
Numerical value that should be formated.}
}
\value{
A charachter string.
}
\author{
Kevin Jakob
}

