\name{get.CALC.RISK.CONT-methods}
\docType{methods}
\alias{get.CALC.RISK.CONT-methods}
\alias{get.CALC.RISK.CONT}
\alias{get.CALC.RISK.CONT,crp.CSFP-method}
\title{Get the value of the \code{CALC.RISK.CONT} flag.}
\description{
The method returns the value of \code{CALC.RISK.CONT} as logical.
}
\keyword{methods}
