\name{get.changes.crp.calc.portfolio.statistics-methods}
\docType{methods}
\alias{get.changes.crp.calc.portfolio.statistics-methods}
\alias{get.changes.crp.calc.portfolio.statistics}
\alias{get.changes.crp.calc.portfolio.statistics,crp.CSFP-method}
\title{Get the state of the \code{changes.crp.calc.portfolio.statistics} flag.}
\description{
The method returns the value of \code{changes.crp.calc.portfolio.statistics}.}
\keyword{methods}
