\name{get.changes.crp.CSFP.loss-methods}
\docType{methods}
\alias{get.changes.crp.CSFP.loss-methods}
\alias{get.changes.crp.CSFP.loss}
\alias{get.changes.crp.CSFP.loss,crp.CSFP-method}
\title{Get the state of the \code{changes.crp.CSFP.loss} flag.}
\description{
The method returns the value of \code{changes.crp.CSFP.loss}.}
\keyword{methods}
