\name{get.changes.crp.CSFP.rc.sd-methods}
\docType{methods}
\alias{get.changes.crp.CSFP.rc.sd-methods}
\alias{get.changes.crp.CSFP.rc.sd}
\alias{get.changes.crp.CSFP.rc.sd,crp.CSFP-method}
\title{Get the state of the \code{changes.crp.CSFP.rc.sd} flag.}
\description{
The method returns the value of \code{changes.crp.CSFP.rc.sd}.}
\keyword{methods}
