\name{get.changes.crp.CSFP.rc.vares-methods}
\docType{methods}
\alias{get.changes.crp.CSFP.rc.vares-methods}
\alias{get.changes.crp.CSFP.rc.vares}
\alias{get.changes.crp.CSFP.rc.vares,crp.CSFP-method}
\title{Get the state of the \code{changes.crp.CSFP.rc.vares}flag.}
\description{
The method returns the value of \code{changes.crp.CSFP.rc.vares}.}
\keyword{methods}
