\name{get.changes.crp.export-methods}
\docType{methods}
\alias{get.changes.crp.export-methods}
\alias{get.changes.crp.export}
\alias{get.changes.crp.export,crp.CSFP-method}
\title{Get the state of the \code{changes.crp.export} flag.}
\description{
The method returns the value of \code{changes.crp.export}.}
\keyword{methods}
