\name{get.changes.crp.measure-methods}
\docType{methods}
\alias{get.changes.crp.measure-methods}
\alias{get.changes.crp.measure}
\alias{get.changes.crp.measure,crp.CSFP-method}
\title{Get the state of the \code{changes.crp.measure} flag.}
\description{
The method returns the value of \code{changes.crp.measure}.}
\keyword{methods}
