\name{get.changes.crp.plausi-methods}
\docType{methods}
\alias{get.changes.crp.plausi-methods}
\alias{get.changes.crp.plausi}
\alias{get.changes.crp.plausi,crp.CSFP-method}
\title{Get the state of the \code{changes.crp.plausi} flag.}
\description{
The method returns the value of \code{changes.crp.plausi}.}
\keyword{methods}
