\name{get.changes.crp.plot-methods}
\docType{methods}
\alias{get.changes.crp.plot-methods}
\alias{get.changes.crp.plot}
\alias{get.changes.crp.plot,crp.CSFP-method}
\title{Get the state of the \code{changes.crp.plot} flag.}
\description{
The method returns the value of \code{changes.crp.plot}.}
\keyword{methods}
