\name{get.changes.crp.read-methods}
\docType{methods}
\alias{get.changes.crp.read-methods}
\alias{get.changes.crp.read}
\alias{get.changes.crp.read,crp.CSFP-method}
\title{Get the state of the \code{changes.crp.read} flag.}
\description{
The method returns the value of \code{changes.crp.read}.
}
\keyword{methods}
