\name{get.CP.RATING-methods}
\docType{methods}
\alias{get.CP.RATING-methods}
\alias{get.CP.RATING}
\alias{get.CP.RATING,crp.CSFP-method}
\title{Get the ratings of the several counterparties of the model}
\description{
The method returns the value of \code{CP.RATING} as numeric.}
\keyword{methods}
