\name{get.ES.TAU.CONT-methods}
\docType{methods}
\alias{get.ES.TAU.CONT-methods}
\alias{get.ES.TAU.CONT}
\alias{get.ES.TAU.CONT,crp.CSFP-method}
\title{Get the corresponding tau for expected shortfall contributions}
\description{
The method returns the value of \code{ES.TAU.CONT} as numeric.}
\keyword{methods}
