\name{get.NITER.MAX-methods}
\docType{methods}
\alias{get.NITER.MAX-methods}
\alias{get.NITER.MAX}
\alias{get.NITER.MAX,crp.CSFP-method}
\title{Get the desired number of iterations or cdf level for loss distribution}
\description{
 The method returns the value of \code{NITER.MAX} as numeric.
}
\keyword{methods}
