\name{get.PDVAR.NAME-methods}
\docType{methods}
\alias{get.PDVAR.NAME-methods}
\alias{get.PDVAR.NAME}
\alias{get.PDVAR.NAME,crp.CSFP-method}
\title{Get the name of the file with the sector variances}
\description{
 The method returns the value of \code{PDVAR.NAME} as character.
}
\keyword{methods}
