\name{get.PLOT.RANGE.Y-methods}
\docType{methods}
\alias{get.PLOT.RANGE.Y-methods}
\alias{get.PLOT.RANGE.Y}
\alias{get.PLOT.RANGE.Y,crp.CSFP-method}
\title{Get the plot range for probabilities}
\description{
 The method returns the value of \code{PLOT.RANGE.Y} as numeric.
}
\keyword{methods}
