\name{get.rating.PD-methods}
\docType{methods}
\alias{get.rating.PD-methods}
\alias{get.rating.PD}
\alias{get.rating.PD,crp.CSFP-method}
\title{Get the probabilities of default corresponding to rating classes}
\description{
The method returns the value of \code{rating.PD} as numeric.}
\keyword{methods}
