\name{get.rating.SD-methods}
\docType{methods}
\alias{get.rating.SD-methods}
\alias{get.rating.SD}
\alias{get.rating.SD,crp.CSFP-method}
\title{Get the standard deviations corresponding to rating classes}
\description{
The method returns the value of \code{rating.SD} as numeric.}
\keyword{methods}
