\name{get.RISK.NAME-methods}
\docType{methods}
\alias{get.RISK.NAME-methods}
\alias{get.RISK.NAME}
\alias{get.RISK.NAME,crp.CSFP-method}
\title{Get the name of the file containing the risk matrix of the model}
\description{
The method returns the value of \code{RISK.NAME} as character.}
\keyword{methods}
