\name{get.SD.CONT-methods}
\docType{methods}
\alias{get.SD.CONT-methods}
\alias{get.SD.CONT}
\alias{get.SD.CONT,crp.CSFP-method}
\title{Get the standard deviation contributions of the several counterparties}
\description{
The method returns the value of \code{SD.CONT} as numeric.}
\keyword{methods}
