\name{get.SEC.VAR.EST-methods}
\docType{methods}
\alias{get.SEC.VAR.EST-methods}
\alias{get.SEC.VAR.EST}
\alias{get.SEC.VAR.EST,crp.CSFP-method}
\title{Get the mode for sector variance estimation of the model}
\description{
The method returns the value of \code{SEC.VAR.EST} as numeric.}
\keyword{methods}
