\name{get.VaR.CONT-methods}
\docType{methods}
\alias{get.VaR.CONT-methods}
\alias{get.VaR.CONT}
\alias{get.VaR.CONT,crp.CSFP-method}
\title{Get the value at risk contributions on counterparty level}
\description{
The method returns the value of \code{VaR.CONT} as numeric.}
\keyword{methods}
