\name{loss}
\docType{methods}
\alias{loss}
\alias{get.LOSS}
\alias{get.LOSS,crp.CSFP-method}
\alias{loss-methods}
\alias{loss,crp.CSFP-method}
\title{Get the several losses (exposure bands) of the model}
\description{
The method returns the value of \code{loss} as numeric.}
\keyword{methods}
