\name{loss.unit}
\docType{methods}
\alias{loss.unit}
\alias{get.LOSS.UNIT}
\alias{get.LOSS.UNIT,crp.CSFP-method}
\alias{loss.unit-methods}
\alias{loss.unit,crp.CSFP-method}
\title{Get the loss unit of the model}
\description{
The method returns the value of \code{loss.unit} as numeric.}
\keyword{methods}
