\name{loss.unit<-}
\docType{methods}
\alias{loss.unit<-}
\alias{set.LOSS.UNIT<-}
\alias{set.LOSS.UNIT<-,crp.CSFP,numeric-method}
\alias{loss.unit<--methods}
\alias{loss.unit<-,crp.CSFP,numeric-method}
\title{Set the loss unit}
\description{
The method changes the models value of \code{loss.unit}}
\keyword{methods}
