\name{measure}
\docType{methods}
\alias{measure}
\alias{crp.measure}
\alias{crp.measure,crp.CSFP-method}
\alias{measure-methods}
\alias{measure,crp.CSFP-method}
\title{Calculating portfolio measures}
\description{
This method calculates key numbers from the discrete loss distribution, such as expected loss \code{EL.crp}, standard deviation \code{SD.crp}, value at risk \code{VaR} and expected shortfall \code{ES} to the given levels in \code{alpha}, prints them on the screen and uses \code{write.summary} to write them - together with the input parameter - to the output directory.
}
\seealso{
\code{\link{crp.CSFP}},
\code{\link{crp.CSFP-class}},
}
\keyword{methods}
