\name{mu.k}
\docType{methods}
\alias{mu.k}
\alias{get.MU.K}
\alias{get.MU.K,crp.CSFP-method}
\alias{mu.k-methods}
\alias{mu.k,crp.CSFP-method}
\title{Get the expected number of defauls per sector}
\description{
 The method returns the value of \code{mu.k} as numeric.
}
\keyword{methods}
