\name{Niter.max}
\docType{methods}
\alias{Niter.max}
\alias{get.NITER.MAX}
\alias{get.NITER.MAX,crp.CSFP-method}
\alias{Niter.max-methods}
\alias{Niter.max,crp.CSFP-method}
\title{Get the desired number of iterations or cdf level for loss distribution}
\description{
 The method returns the value of \code{Niter.max} as numeric.
}
\keyword{methods}
