\name{Niter.max<-}
\docType{methods}
\alias{Niter.max<-}
\alias{set.NITER.MAX<-}
\alias{set.NITER.MAX<-,crp.CSFP,numeric-method}
\alias{Niter.max<--methods}
\alias{Niter.max<-,crp.CSFP,numeric-method}
\title{Set the maximal number of iterations or desired cdf level}
\description{
The method changes the models value of \code{Niter.max}}
\keyword{methods}
