\name{path.in}
\docType{methods}
\alias{path.in}
\alias{get.PATH.IN}
\alias{get.PATH.IN,crp.CSFP-method}
\alias{path.in-methods}
\alias{path.in,crp.CSFP-method}
\title{Get the input path of the model}
\description{
 The method returns the value of \code{path.in} as character.
}
\keyword{methods}
