\name{path.in<-}
\docType{methods}
\alias{path.in<-}
\alias{set.PATH.IN<-}
\alias{set.PATH.IN<-,crp.CSFP,character-method}
\alias{path.in<--methods}
\alias{path.in<-,crp.CSFP,character-method}
\title{Set input path}
\description{
The method changes the models value of \code{path.in}}
\keyword{methods}