\name{path.out}
\docType{methods}
\alias{path.out}
\alias{get.PATH.OUT}
\alias{get.PATH.OUT,crp.CSFP-method}
\alias{path.out-methods}
\alias{path.out,crp.CSFP-method}
\title{Get the output path of the model}
\description{
 The method returns the value of \code{path.out} as character.
}
\keyword{methods}
