\name{path.out<-}
\docType{methods}
\alias{path.out<-}
\alias{set.PATH.OUT<-}
\alias{set.PATH.OUT<-,crp.CSFP,character-method}
\alias{path.out<--methods}
\alias{path.out<-,crp.CSFP,character-method}
\title{Set output path}
\description{
The method changes the models value of \code{path.out}}
\keyword{methods}
