\name{PD}
\docType{methods}
\alias{PD}
\alias{get.PD0}
\alias{get.PD0,crp.CSFP-method}
\alias{PD-methods}
\alias{PD,crp.CSFP-method}
\title{Get the counterparty probabilities of default of the model}
\description{
 The method returns the value of \code{PD} as numeric.
}
\keyword{methods}
