\name{PD.crp}
\docType{methods}
\alias{PD.crp}
\alias{get.PD}
\alias{get.PD,crp.CSFP-method}
\alias{PD.crp-methods}
\alias{PD.crp,crp.CSFP-method}
\title{Get the counterparty probabilities of default after discretization}
\description{
 The method returns the value of \code{PD.crp} as numeric.
}
\keyword{methods}
