\name{pd_sector_var}
\docType{data}
\alias{pd_sector_var}
\title{Sector variances for the Credit Suisse example portfolio}
\description{
  The file contains the sector variances corrsponding to the example portfolio from the Credit Suisse First Boston for CreditRisk+.
}
\references{First Boston Financial Products, "CreditRisk+", 1997}
\keyword{datasets}