\name{PL.crp}
\docType{methods}
\alias{PL.crp}
\alias{get.PL}
\alias{get.PL,crp.CSFP-method}
\alias{PL.crp-methods}
\alias{PL.crp,crp.CSFP-method}
\title{Get the potetnial losses per counterparty after discretization}
\description{
 The method returns the value of \code{PL.crp} as numeric.
}
\keyword{methods}

