\name{plot}
\docType{methods}
\alias{plot}
\alias{crp.plot}
\alias{crp.plot,crp.CSFP-method}
\alias{plot-methods}
\alias{plot,crp.CSFP-method}
\title{Plotting the PDF}
\description{
This method creates a plot in a new window of the loss distribution together with lines indicating the \code{EL.crp}, \code{VaR} and \code{ES}. The X-/Y- ranges are taken from \code{plot.range.x /-y}.
}
\seealso{
\code{\link{crp.CSFP}},
\code{\link{crp.CSFP-class}},
}
\keyword{methods}
