\name{PLOT.PDF}
\docType{methods}
\alias{PLOT.PDF}
\alias{get.PLOT.PDF}
\alias{get.PLOT.PDF,crp.CSFP-method}
\alias{PLOT.PDF-methods}
\alias{PLOT.PDF,crp.CSFP-method}
\title{Get the state of \code{PLOT.PDF}}
\description{
 The method returns the value of \code{PLOT.PDF} as logical.
}
\keyword{methods}
