\name{PLOT.PDF<-}
\docType{methods}
\alias{PLOT.PDF<-}
\alias{set.PLOT.PDF<-}
\alias{set.PLOT.PDF<-,crp.CSFP,logical-method}
\alias{PLOT.PDF<--methods}
\alias{PLOT.PDF<-,crp.CSFP,logical-method}
\title{Set the state of \code{PLOT.PDF}}
\description{
The method changes the models value of \code{PLOT.PDF}}
\keyword{methods}
