\name{PLOT.range.x}
\docType{methods}
\alias{PLOT.range.x}
\alias{get.PLOT.RANGE.X}
\alias{get.PLOT.RANGE.X,crp.CSFP-method}
\alias{PLOT.range.x-methods}
\alias{PLOT.range.x,crp.CSFP-method}
\title{Get the plot range for losses}
\description{
 The method returns the value of \code{PLOT.range.x} as numeric.
}
\keyword{methods}
