\name{PLOT.range.x<-}
\docType{methods}
\alias{PLOT.range.x<-}
\alias{set.PLOT.RANGE.X<-}
\alias{set.PLOT.RANGE.X<-,crp.CSFP,numeric-method}
\alias{PLOT.range.x<--methods}
\alias{PLOT.range.x<-,crp.CSFP,numeric-method}
\title{Set the plot range for the losses}
\description{
The method changes the models value of \code{PLOT.range.x}}
\keyword{methods}
