\name{PLOT.range.y}
\docType{methods}
\alias{PLOT.range.y}
\alias{get.PLOT.RANGE.Y}
\alias{get.PLOT.RANGE.Y,crp.CSFP-method}
\alias{PLOT.range.y-methods}
\alias{PLOT.range.y,crp.CSFP-method}
\title{Get the plot range for probabilities}
\description{
 The method returns the value of \code{PLOT.range.y} as numeric.
}
\keyword{methods}
