\name{PLOT.range.y<-}
\docType{methods}
\alias{PLOT.range.y<-}
\alias{set.PLOT.RANGE.Y<-}
\alias{set.PLOT.RANGE.Y<-,crp.CSFP,numeric-method}
\alias{PLOT.range.y<--methods}
\alias{PLOT.range.y<-,crp.CSFP,numeric-method}
\title{Set the plot range for the probabilities}
\description{
The method changes the models value of \code{PLOT.range.y}}
\keyword{methods}

