\name{PLOT.scale}
\docType{methods}
\alias{PLOT.scale}
\alias{get.PLOT.SCALE}
\alias{get.PLOT.SCALE,crp.CSFP-method}
\alias{PLOT.scale-methods}
\alias{PLOT.scale,crp.CSFP-method}
\title{Get the plot scale for losses}
\description{
 The method returns the value of \code{PLOT.scale} as numeric.
}
\keyword{methods}

