\name{PLOT.scale<-}
\docType{methods}
\alias{PLOT.scale<-}
\alias{set.PLOT.SCALE<-}
\alias{set.PLOT.SCALE<-,crp.CSFP,numeric-method}
\alias{PLOT.scale<--methods}
\alias{PLOT.scale<-,crp.CSFP,numeric-method}
\title{Set the plot scale for portfolio losses}
\description{
The method changes the models value of \code{PLOT.scale}}
\keyword{methods}
