\name{port.name}
\docType{methods}
\alias{port.name}
\alias{get.PORT.NAME}
\alias{get.PORT.NAME,crp.CSFP-method}
\alias{port.name-methods}
\alias{port.name,crp.CSFP-method}
\title{Get the name of the portfolio file}
\description{
 The method returns the value of \code{port.name} as character.
}
\keyword{methods}
