\name{port.name<-}
\docType{methods}
\alias{port.name<-}
\alias{set.PORT.NAME<-}
\alias{set.PORT.NAME<-,crp.CSFP,character-method}
\alias{port.name<--methods}
\alias{port.name<-,crp.CSFP,character-method}
\title{Set the name for the portfolio file}
\description{
The method changes the models value of \code{port.name}}
\keyword{methods}
