\name{portfolio}
\docType{data}
\alias{portfolio}
\title{Portfolio data for the Credit Suisse example portfolio}
\description{
The file contains the portfolio corrsponding to the example portfolio from the Credit Suisse First Boston for CreditRisk+.
}
\references{First Boston Financial Products, "CreditRisk+", 1997}
\keyword{datasets}