\name{rating}
\docType{methods}
\alias{rating}
\alias{get.rating}
\alias{get.rating,crp.CSFP-method}
\alias{rating-methods}
\alias{rating,crp.CSFP-method}
\title{Get the rating classes of the model}
\description{
The method returns the value of \code{rating} as numeric.}
\keyword{methods}
