\name{rating.PD}
\docType{methods}
\alias{rating.PD}
\alias{get.rating.PD}
\alias{get.rating.PD,crp.CSFP-method}
\alias{rating.PD-methods}
\alias{rating.PD,crp.CSFP-method}
\title{Get the PDs of rating classes}
\description{
The method returns the value of \code{rating.PD} as numeric.}
\section{Methods}{
\describe{

\item{\code{signature(this = "crp.CSFP")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
