\name{rating.scale.name}
\docType{methods}
\alias{rating.scale.name}
\alias{get.RISK.NAME}
\alias{get.RISK.NAME,crp.CSFP-method}
\alias{rating.scale.name-methods}
\alias{rating.scale.name,crp.CSFP-method}
\title{Get the name of the file containing the risk matrix of the model}
\description{
The method returns the value of \code{rating.scale.name} as character.}
\keyword{methods}
