\name{rating.scale.name<-}
\docType{methods}
\alias{rating.scale.name<-}
\alias{set.RISK.NAME<-}
\alias{set.RISK.NAME<-,crp.CSFP,character-method}
\alias{rating.scale.name<--methods}
\alias{rating.scale.name<-,crp.CSFP,character-method}
\title{Set the name for the file containing the rating scale}
\description{
The method changes the models value of \code{rating.scale.name}}
\keyword{methods}
