\name{rating_pd}
\docType{data}
\alias{rating_pd}
\title{Risk matrix for the Credit Suisse example portfolio}
\description{
  The file contains the risk matrix corrsponding to the example portfolio from the Credit Suisse First Boston for CreditRisk+.
}
\references{First Boston Financial Products, "CreditRisk+", 1997}
\keyword{datasets}