\name{rc.sd}
\docType{methods}
\alias{rc.sd}
\alias{crp.CSFP.rc.sd}
\alias{crp.CSFP.rc.sd,crp.CSFP-method}
\alias{rc.sd-methods}
\alias{rc.sd,crp.CSFP-method}
\title{Calculating risk contributions to standard deviation}
\description{
This method calculates the contributions on counterparty level to the portfolio standard deviation. No data out of \code{loss.dist} are required.
}
\seealso{
\code{\link{rc.vares}},
\code{\link{export}},
\code{\link{crp.CSFP-class}},
\code{\link{crp.CSFP}},
}
\keyword{methods}
