\name{save.memory}
\docType{methods}
\alias{save.memory}
\alias{get.save.memory}
\alias{get.save.memory,crp.CSFP-method}
\alias{save.memory-methods}
\alias{save.memory,crp.CSFP-method}
\title{Get the state of \code{save.memory}}
\description{
The method returns the value of \code{save.memory} as logical.}
\keyword{methods}