\name{save.memory<-}
\docType{methods}
\alias{save.memory<-}
\alias{set.save.memory<-}
\alias{set.save.memory<-,crp.CSFP,logical-method}
\alias{save.memory<--methods}
\alias{save.memory<-,crp.CSFP,logical-method}
\title{Set the state of \code{save.memory}}
\description{
The method changes the models value of \code{save.memory}}
\keyword{methods}
