\name{SD.cont}
\docType{methods}
\alias{SD.cont}
\alias{get.SD.CONT}
\alias{get.SD.CONT,crp.CSFP-method}
\alias{SD.cont-methods}
\alias{SD.cont,crp.CSFP-method}
\title{Get the contributions to standard deviation}
\description{
The method returns the value of \code{SD.cont} as numeric.}
\keyword{methods}
