\name{SD.crp}
\docType{methods}
\alias{SD.crp}
\alias{get.SD.crp}
\alias{get.SD.crp,crp.CSFP-method}
\alias{SD.crp-methods}
\alias{SD.crp,crp.CSFP-method}
\title{Get the discretized standard deviation of loss distribution}
\description{
The method returns the value of \code{SD.crp} as numeric.}
\keyword{methods}
