\name{sec.var}
\docType{methods}
\alias{sec.var}
\alias{get.SEC.VAR}
\alias{get.SEC.VAR,crp.CSFP-method}
\alias{sec.var-methods}
\alias{sec.var,crp.CSFP-method}
\title{Get self estimated sector variances}
\description{
The method returns the value of \code{sec.var} as numeric.}
\keyword{methods}
