\name{sec.var.est}
\docType{methods}
\alias{sec.var.est}
\alias{get.SEC.VAR.EST}
\alias{get.SEC.VAR.EST,crp.CSFP-method}
\alias{sec.var.est-methods}
\alias{sec.var.est,crp.CSFP-method}
\title{Get the mode for sector variance estimation}
\description{
The method returns the models value of \code{sec.var.est}}
\keyword{methods}
