\name{sec.var.est<-}
\docType{methods}
\alias{sec.var.est<-}
\alias{set.SEC.VAR.EST<-}
\alias{set.SEC.VAR.EST<-,crp.CSFP,numeric-method}
\alias{sec.var.est<--methods}
\alias{sec.var.est<-,crp.CSFP,numeric-method}
\title{Set the mode for sector variance estimation}
\description{
The method changes the models value of \code{sec.var.est}}
\keyword{methods}
