\name{sec.var.name}
\docType{methods}
\alias{sec.var.name}
\alias{get.PDVAR.NAME}
\alias{get.PDVAR.NAME,crp.CSFP-method}
\alias{sec.var.name-methods}
\alias{sec.var.name,crp.CSFP-method}
\title{Set the name of the file with the sector variances}
\description{
 The method changes the value of \code{sec.var.name}.
}
\keyword{methods}
