\name{sec.var.name<-}
\docType{methods}
\alias{sec.var.name<-}
\alias{set.PDVAR.NAME<-}
\alias{set.PDVAR.NAME<-,crp.CSFP,character-method}
\alias{sec.var.name<--methods}
\alias{sec.var.name<-,crp.CSFP,character-method}
\title{Set the name of the file with the sector variances}
\description{
 The method changes the value of \code{sec.var.name}.
}
\keyword{methods}
