\name{set.CALC.RISK.CONT<--methods}
\docType{methods}
\alias{set.CALC.RISK.CONT<--methods}
\alias{set.CALC.RISK.CONT<-}
\alias{set.CALC.RISK.CONT<-,crp.CSFP,logical-method}
\title{Set the state of \code{CALC.RISK.CONT}.}
\description{
The method changes the models value of \code{CALC.RISK.CONT} to the value after '<-'}
\keyword{methods}
