\name{set.file.format<--methods}
\docType{methods}
\alias{set.file.format<--methods}
\alias{set.file.format<-}
\alias{set.file.format<-,crp.CSFP,character-method}
\title{Set the file format}
\description{
The method changes the models value of \code{file.format} to the value after '<-'}
\keyword{methods}
