\name{set.LOSS.UNIT<--methods}
\docType{methods}
\alias{set.LOSS.UNIT<--methods}
\alias{set.LOSS.UNIT<-}
\alias{set.LOSS.UNIT<-,crp.CSFP,numeric-method}
\title{Set the loss unit}
\description{
The method changes the models value of \code{LOSS.UNIT} to the value after '<-'}
\keyword{methods}
