\name{set.NITER.MAX<--methods}
\docType{methods}
\alias{set.NITER.MAX<--methods}
\alias{set.NITER.MAX<-}
\alias{set.NITER.MAX<-,crp.CSFP,numeric-method}
\title{Set the maximal number of iterations or desired cdf level}
\description{
The method changes the models value of \code{NITER.MAX} to the value after '<-'}
\keyword{methods}
