\name{set.PDVAR.NAME<--methods}
\docType{methods}
\alias{set.PDVAR.NAME<--methods}
\alias{set.PDVAR.NAME<-}
\alias{set.PDVAR.NAME<-,crp.CSFP,character-method}
\title{Set the name for the file with sector variances}
\description{
The method changes the models value of \code{PDVAR.NAME} to the value after '<-'}
\keyword{methods}
