\name{set.PLOT.PDF<--methods}
\docType{methods}
\alias{set.PLOT.PDF<--methods}
\alias{set.PLOT.PDF<-}
\alias{set.PLOT.PDF<-,crp.CSFP,logical-method}
\title{Set the state of \code{PLOT.PDF}}
\description{
The method changes the models value of \code{PLOT.PDF} to the value after '<-'}
\keyword{methods}
