\name{set.PLOT.RANGE.X<--methods}
\docType{methods}
\alias{set.PLOT.RANGE.X<--methods}
\alias{set.PLOT.RANGE.X<-}
\alias{set.PLOT.RANGE.X<-,crp.CSFP,numeric-method}
\title{Set the plot range for the losses}
\description{
The method changes the models value of \code{PLOT.RANGE.X} to the value after '<-'}
\keyword{methods}
