\name{set.PLOT.RANGE.Y<--methods}
\docType{methods}
\alias{set.PLOT.RANGE.Y<--methods}
\alias{set.PLOT.RANGE.Y<-}
\alias{set.PLOT.RANGE.Y<-,crp.CSFP,numeric-method}
\title{Set the plot range for the probabilities}
\description{
The method changes the models value of \code{PLOT.RANGE.Y} to the value after '<-'}
\keyword{methods}
