\name{set.PLOT.SCALE<--methods}
\docType{methods}
\alias{set.PLOT.SCALE<--methods}
\alias{set.PLOT.SCALE<-}
\alias{set.PLOT.SCALE<-,crp.CSFP,numeric-method}
\title{Set the plot scale for portfolio losses}
\description{
The method changes the models value of \code{PLOT.SCALE} to the value after '<-'}
\keyword{methods}
