\name{set.PORT.NAME<--methods}
\docType{methods}
\alias{set.PORT.NAME<--methods}
\alias{set.PORT.NAME<-}
\alias{set.PORT.NAME<-,crp.CSFP,character-method}
\title{Set the name for the portfolio file}
\description{
The method changes the models value of \code{PORT.NAME} to the value after '<-'}
\keyword{methods}
