\name{set.RISK.NAME<--methods}
\docType{methods}
\alias{set.RISK.NAME<--methods}
\alias{set.RISK.NAME<-}
\alias{set.RISK.NAME<-,crp.CSFP,character-method}
\title{Set the name for the file with the risk matrix}
\description{
The method changes the models value of \code{RISK.NAME} to the value after '<-'}
\keyword{methods}
