\name{set.save.memory<--methods}
\docType{methods}
\alias{set.save.memory<--methods}
\alias{set.save.memory<-}
\alias{set.save.memory<-,crp.CSFP,logical-method}
\title{Set the state of \code{save.memory}}
\description{
The method changes the models value of \code{save.memory} to the value after '<-'}
\keyword{methods}
