\name{set.SEC.VAR.EST<--methods}
\docType{methods}
\alias{set.SEC.VAR.EST<--methods}
\alias{set.SEC.VAR.EST<-}
\alias{set.SEC.VAR.EST<-,crp.CSFP,numeric-method}
\title{Set the mode for sector variance estimation}
\description{
The method changes the models value of \code{SEC.VAR.EST} to the value after '<-'}
\keyword{methods}
