\name{show}
\docType{methods}
\alias{show}
\alias{show-methods}
\alias{show,crp.CSFP-method}
\title{Show summary of object crp.CSFP}
\description{
This method extends the show function for objects of class \code{crp.CSFP}. A short summary with the most important model parameters is printed.
}
\seealso{
\code{\link{show}}
\code{\link{crp.CSFP}},
\code{\link{crp.CSFP-class}},
}
\keyword{methods}
\keyword{show}
