\name{show.crp.CSFP}
\docType{methods}
\alias{show.crp.CSFP}
\title{show method for crp.CSFP class}
\description{
This method extends the show function for objects of class \code{crp.CSFP}. A short summary with the most important model parameters is printed.
}
\seealso{
\code{\link{show}}
\code{\link{crp.CSFP}},
\code{\link{crp.CSFP-class}},
}
\keyword{methods}
\keyword{show}
