\name{sigma.k}
\docType{methods}
\alias{sigma.k}
\alias{get.SIGMA.K}
\alias{get.SIGMA.K,crp.CSFP-method}
\alias{sigma.k-methods}
\alias{sigma.k,crp.CSFP-method}
\title{Get the sector standard deviation}
\description{
The method returns the value of \code{sigma.k} as numeric.}
\keyword{methods}