\name{sigma.sqr.div}
\docType{methods}
\alias{sigma.sqr.div}
\alias{get.SIGMA2_DIV}
\alias{get.SIGMA2_DIV,crp.CSFP-method}
\alias{sigma.sqr.div-methods}
\alias{sigma.sqr.div,crp.CSFP-method}
\title{Get the diversifiable risk of the model}
\description{
The method returns the value of \code{sigma.sqr.div} (already squared) as numeric.}
\keyword{methods}
