\name{sigma.sqr.syst}
\docType{methods}
\alias{sigma.sqr.syst}
\alias{get.SIGMA2_SYST}
\alias{get.SIGMA2_SYST,crp.CSFP-method}
\alias{sigma.sqr.syst-methods}
\alias{sigma.sqr.syst,crp.CSFP-method}
\title{Get the systematik risk of the model}
\description{
The method returns the value of \code{sigma.sqr.syst} (already squared) as numeric.}
\keyword{methods}
