\name{sigma_k}
\docType{methods}
\alias{sigma.k}
\alias{get.SIGMA.K}
\alias{get.SIGMA.K,crp.CSFP-method}
\alias{sigma.k-methods}
\alias{sigma.k,crp.CSFP-method}
\alias{sigma_k}
\alias{sigma_k-methods}
\alias{sigma_k,crp.CSFP-method}

\title{Get the sector standard deviation}
\description{
The method returns the value of \code{sigma_k} as numeric.
The variable/method was renamed from \code{sigma.k} to \code{sigma_k} for S3 compatibility.}
\keyword{methods}
