\name{sigma_sqr_div}
\docType{methods}
\alias{sigma.sqr.div}
\alias{get.SIGMA2_DIV}
\alias{get.SIGMA2_DIV,crp.CSFP-method}
\alias{sigma.sqr.div-methods}
\alias{sigma.sqr.div,crp.CSFP-method}
\alias{sigma_sqr_div}
\alias{sigma_sqr_div-methods}
\alias{sigma_sqr_div,crp.CSFP-method}

\title{Get the diversifiable risk of the model}
\description{
The method returns the value of \code{sigma_sqr_div} (already squared) as numeric.
The variable/method was renamed from \code{sigma.sqr.div} to \code{sigma_sqr_div} for S3 compatibility.}
\keyword{methods}
