\name{sigma.sqr.syst}
\docType{methods}
\alias{sigma.sqr.syst}
\alias{get.SIGMA2_SYST}
\alias{get.SIGMA2_SYST,crp.CSFP-method}
\alias{sigma.sqr.syst-methods}
\alias{sigma.sqr.syst,crp.CSFP-method}
\alias{sigma_sqr_syst}
\alias{sigma_sqr_syst-methods}
\alias{sigma_sqr_syst,crp.CSFP-method}
\title{Get the systematik risk of the model}
\description{
The method returns the value of \code{sigma_sqr_syst} (already squared) as numeric.
The variable/method was renamed from \code{sigma.sqr.syst} to \code{sigma_sqr_syst} for S3 compatibility.}
\keyword{methods}
