\name{summary}
\docType{methods}
\alias{summary}
\alias{crp.summary}
\alias{crp.summary,crp.CSFP-method}
\alias{summary-methods}
\alias{summary,crp.CSFP-method}
\title{Summarize portfolio key numbers}
\description{
This method summarizes the input parameter, portfolio statistics from \code{calc.portfolio.statistics} and the results of \code{measure} in a list.
}
\seealso{
\code{\link{write.summary}},
\code{\link{crp.CSFP}},
\code{\link{crp.CSFP-class}},
}
\keyword{methods}
