\name{VaR.cont}
\docType{methods}
\alias{VaR.cont}
\alias{get.VaR.CONT}
\alias{get.VaR.CONT,crp.CSFP-method}
\alias{VaR.cont-methods}
\alias{VaR.cont,crp.CSFP-method}
\title{Get the value at risk contributions on counterparty level}
\description{
The method returns the value of \code{VaR.cont} as numeric.}
\keyword{methods}
