\name{VaR.pos}
\docType{methods}
\alias{VaR.pos}
\alias{get.SELV}
\alias{get.SELV,crp.CSFP-method}
\alias{VaR.pos-methods}
\alias{VaR.pos,crp.CSFP-method}
\title{Get the position of value at risk in CDF}
\description{
The method returns the value of \code{VaR.pos} as numeric.}
\keyword{methods}
