\name{write.summary}
\docType{methods}
\alias{write.summary}
\alias{write.summary-methods}
\alias{write.summary,crp.CSFP-method}
\alias{crp.write.summary}
\alias{crp.write.summary-methods}
\alias{crp.write.summary,crp.CSFP-method}
\title{Writing summary to file}
\description{
This method writes the list, created by \code{crp.summary} into a file, called "summary.csv" into the output directory.
}
\seealso{
\code{\link{summary}},
\code{\link{crp.CSFP}},
\code{\link{crp.CSFP-class}},
}
\keyword{methods}
